updated math functions to use the sdt version

+++ DebugUtils/Source/DebugDraw.cpp
@@ -383,7 +383,7 @@ inline void vcross(float* dest, const float* v1, const float* v2)
 
 inline void vnormalize(float* v)
 {
-    const float d = 1.0f / sqrtf(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+    const float d = 1.0f / std::sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
     v[0] *= d;
     v[1] *= d;
     v[2] *= d;
@@ -440,7 +440,7 @@ void duAppendArc(duDebugDraw* dd, const float x0, const float y0, const float z0
     const float dx = x1 - x0;
     const float dy = y1 - y0;
     const float dz = z1 - z0;
-    const float len = sqrtf(dx * dx + dy * dy + dz * dz);
+    const float len = std::sqrt(dx * dx + dy * dy + dz * dz);
     float prev[3];
     evalArc(x0, y0, z0, dx, dy, dz, len * h, PAD, prev);
     for (int i = 1; i <= NUM_ARC_PTS; ++i)
